requirements.txt
django==4.2.3
dj-database-url
dj_static
python-decouple
psycopg2-binary>=2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers
django-templated-mail
djoser>=2.1.0
mysqlclient
PyJWT
whitenoise
gunicorn

# can use env file
python-dotenv
